Add files via upload
Abstract — The project aims at enhancing logistics by optimizing operational efficiency through simulation for 'We-Doo', a recent startup company. The primary objective of this project is to recommend an optimal delivery warehouse location in the commuter township map with the lowest cost base. The methodology for this project involves developing a simulation model to evaluate candidate warehouse locations through heuristics by incorporating factors such as customer location, route length, and operational cost of the delivery. The workload data is generated by using the seed number 8268 to simulate daily delivery operations, considering parameters such as length of delivery route, count of parcels delivered, and working time of the driver. Through thorough investigation, analysis, and simulation runs for multiple days, valuable insights are obtained regarding the performance of different warehouse locations and their respective cost base. Evaluation of the simulation involves histograms for visualisation and statistical significance is investigated by using ANOVA. The sample size was increased to 50 days in the longer simulation run to discover valuable insights for strategic decision making, which highlight the significance of strategic location selection in enhancing operational efficiency and customer satisfaction.